services/ros/requirements-ros.txt
fastapi
uvicorn[standard]
aiokafka
pydantic
httpx
# Optional Firestore support:
google-cloud-firestore
